Marketer service and bug fix

+++ redpraisal-api/src/main/java/io/github/grandlich/redpraisal/dto/AppraisalResponseItemPricesSection.java
@@ -2,7 +2,6 @@ package io.github.grandlich.redpraisal.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import lombok.Data;
 
 @Data
@@ -10,9 +9,9 @@ import lombok.Data;
 public class AppraisalResponseItemPricesSection {
 
   private BigDecimal avg;
-  private BigInteger max;
-  private BigInteger median;
-  private BigInteger min;
+  private BigDecimal max;
+  private BigDecimal median;
+  private BigDecimal min;
   private Long order_count;
   private BigDecimal percentile;
   private Long stddev;
